# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "noviy_2026_bot API",
    "version": "0.1.0"
  },
  "paths": {
    "/prizes": {
      "get": {
        "tags": [
          "prizes"
        ],
        "summary": "List Prizes",
        "operationId": "list_prizes_prizes_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PrizeOut"
                  },
                  "title": "Response List Prizes Prizes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "prizes"
        ],
        "summary": "Create Prize",
        "operationId": "create_prize_prizes_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrizeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrizeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/prizes/remaining": {
      "get": {
        "tags": [
          "prizes"
        ],
        "summary": "List Remaining",
        "operationId": "list_remaining_prizes_remaining_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 500,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PrizeRemainingOut"
                  },
                  "title": "Response List Remaining Prizes Remaining Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/prizes/remaining/{prize_id}": {
      "get": {
        "tags": [
          "prizes"
        ],
        "summary": "Get Remaining",
        "operationId": "get_remaining_prizes_remaining__prize_id__get",
        "parameters": [
          {
            "name": "prize_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prize Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrizeRemainingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "prizes"
        ],
        "summary": "Upsert Remaining",
        "operationId": "upsert_remaining_prizes_remaining__prize_id__put",
        "parameters": [
          {
            "name": "prize_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prize Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrizeRemainingUpsert"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrizeRemainingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "prizes"
        ],
        "summary": "Delete Remaining",
        "operationId": "delete_remaining_prizes_remaining__prize_id__delete",
        "parameters": [
          {
            "name": "prize_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prize Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/prizes/wins": {
      "get": {
        "tags": [
          "prizes"
        ],
        "summary": "List Wins",
        "operationId": "list_wins_prizes_wins_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PrizeWinOut"
                  },
                  "title": "Response List Wins Prizes Wins Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/prizes/wins/by-user/{user_id}": {
      "get": {
        "tags": [
          "prizes"
        ],
        "summary": "List Wins By User",
        "operationId": "list_wins_by_user_prizes_wins_by_user__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PrizeWinOut"
                  },
                  "title": "Response List Wins By User Prizes Wins By User  User Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/prizes/draw": {
      "post": {
        "tags": [
          "prizes"
        ],
        "summary": "Draw Prize",
        "operationId": "draw_prize_prizes_draw_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrizeDrawIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrizeDrawOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/prizes/{prize_id}": {
      "get": {
        "tags": [
          "prizes"
        ],
        "summary": "Get Prize",
        "operationId": "get_prize_prizes__prize_id__get",
        "parameters": [
          {
            "name": "prize_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prize Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrizeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "prizes"
        ],
        "summary": "Update Prize",
        "operationId": "update_prize_prizes__prize_id__put",
        "parameters": [
          {
            "name": "prize_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prize Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrizeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "prizes"
        ],
        "summary": "Delete Prize",
        "operationId": "delete_prize_prizes__prize_id__delete",
        "parameters": [
          {
            "name": "prize_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prize Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeletedOut": {
        "properties": {
          "deleted": {
            "type": "integer",
            "title": "Deleted"
          }
        },
        "type": "object",
        "required": [
          "deleted"
        ],
        "title": "DeletedOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PrizeCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "friendly_name": {
            "type": "string",
            "title": "Friendly Name"
          },
          "weight": {
            "type": "number",
            "title": "Weight"
          }
        },
        "type": "object",
        "required": [
          "name",
          "friendly_name",
          "weight"
        ],
        "title": "PrizeCreate"
      },
      "PrizeDrawIn": {
        "properties": {
          "voucher": {
            "type": "string",
            "title": "Voucher"
          }
        },
        "type": "object",
        "required": [
          "voucher"
        ],
        "title": "PrizeDrawIn"
      },
      "PrizeDrawOut": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "prize": {
            "$ref": "#/components/schemas/PrizeOut"
          },
          "won_at": {
            "type": "string",
            "format": "date-time",
            "title": "Won At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "prize",
          "won_at"
        ],
        "title": "PrizeDrawOut"
      },
      "PrizeOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "friendly_name": {
            "type": "string",
            "title": "Friendly Name"
          },
          "weight": {
            "type": "number",
            "title": "Weight"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "friendly_name",
          "weight"
        ],
        "title": "PrizeOut"
      },
      "PrizeRemainingOut": {
        "properties": {
          "prize_id": {
            "type": "integer",
            "title": "Prize Id"
          },
          "remaining": {
            "type": "integer",
            "title": "Remaining"
          }
        },
        "type": "object",
        "required": [
          "prize_id",
          "remaining"
        ],
        "title": "PrizeRemainingOut"
      },
      "PrizeRemainingUpsert": {
        "properties": {
          "remaining": {
            "type": "integer",
            "title": "Remaining"
          }
        },
        "type": "object",
        "required": [
          "remaining"
        ],
        "title": "PrizeRemainingUpsert"
      },
      "PrizeWinOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "prize_id": {
            "type": "integer",
            "title": "Prize Id"
          },
          "won_at": {
            "type": "string",
            "format": "date-time",
            "title": "Won At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "prize_id",
          "won_at"
        ],
        "title": "PrizeWinOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}